requirements.txt
aniso8601==8.0.0
apipkg==1.5
asgiref==3.2.3
atomicwrites==1.4.0
attrs==21.2.0
cachelib==0.10.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
configparser==5.0.2
crayons==0.4.0
execnet==1.9.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask==2.2.3
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==8.9.0
packaging==21.0
pip==23.0.1
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
PyJWT==2.6.0
pyparsing==2.4.7
pytest-forked==1.3.0
pytest-rerunfailures==10.1
pytest-xdist==2.3.0
pytest==6.2.5
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
setuptools==41.2.0
six==1.16.0
sqlparse==0.3.0
toml==0.10.2
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.2.3
zipp==3.5.0
Flask-SQLAlchemy==3.0.3
SQLAlchemy~=2.0.7